modified Rotation_Matrix
