+++ Source/Solver/Transport/NEGF/Make.package
@@ -1,6 +1,9 @@
 CEXE_sources += IntegrationPath.cpp
 CEXE_headers += IntegrationPath.H
 
+CEXE_sources += Rotation_Matrix.cpp
+CEXE_headers += Rotation_Matrix.H
+
 CEXE_headers += Matrix_Block_Util.H
 CEXE_headers += Matrix_Block.H
