+++ Source/Solver/Transport/NEGF/Rotation_Matrix.H
@@ -1,102 +1,49 @@
 #ifndef ROTATION_MATRIX_H_
 #define ROTATION_MATRIX_H_
 
-#include "matrix.hpp"
-#include "nvector.hpp"
-#include "ops.hpp"
-#include "rotationMatrix.hpp"
-#include <cmath>
+#include <AMReX_REAL.H>
+#include <AMReX_Vector.H>
 
 enum class AngleType : int { Degrees, Radians };
-
 enum class AxisType : int { X, Y, Z };
 
 template<typename T>
 class c_RotationMatrix
 {
     /*X,Y,Z rotation angles*/
-    amrex::Vector<T> _angles; 
+    amrex::Vector<T> _angles;
     amrex::Vector<amrex::Vector<T>> _rotX;
     amrex::Vector<amrex::Vector<T>> _rotY;
     amrex::Vector<amrex::Vector<T>> _rotZ;
 
-    void Define_RotationMatrices() 
-    {
-        _rotX[0] = {1, 0, 0};
-        _rotX[1] = {0,  cos(_angles[0]), sin(_angles[0])};
-        _rotX[2] = {0, -sin(_angles[0]), cos(_angles[0])};
+    void Define_RotationMatrices();
 
-        _rotY[0] = {cos(_angles[1]), 0, -sin(_angles[1])};
-        _rotY[1] = {0, 1, 0};
-        _rotY[2] = {sin(_angles[1]), 0, cos(_angles[1])};
+    void Set_RotationAngles(amrex::Vector<T>& new_angles);
 
-        _rotZ[0] = { cos(_angles[2]), sin(_angles[2]), 0};
-        _rotZ[1] = {-sin(_angles[2]), cos(_angles[2]), 0};
-        _rotZ[2] = {0, 0, 1};
-    }
+    template<typename ContainerType>
+    void Rotate(ContainerType& v2,
+                const ContainerType v1,
+                const amrex::Vector<amrex::Vector<T>>& rotM);
 
-    void Rotate(amrex::Vector<T>& v2,
-                const amrex::Vector<T> v1, 
-                const amrex::Vector<amrex::Vector<T>>& rotM) 
-    {
-        assert(v1.size() == 3);
-        assert(v2.size() == 3);
-
-        for(int i=0; i<3; ++i) {
-            for(int j=0; j<3; ++j) {
-                v2[i] += rotM[i][j] * v1[j];
-            }
-        }
-    }
- 
-    void ApplyRotationOrder(amrex::Vector<T>& v,
-                            const std::vector<AxisType>& order)
-    {
-        for (AxisType axis : order) 
-        {
-            switch(axis) {
-                case AxisType::X:
-                    Rotate(v, v, _rotX);
-                    break;
-                case AxisType::Y:
-                    Rotate(v, v, _rotY);
-                    break;
-                case AxisType::Z:
-                    Rotate(v, v, _rotZ);
-                    break;
-            }
-        }
-    }    
+    template<typename ContainerType>
+    void Apply_RotationOrder(ContainerType& v,
+                             const std::vector<AxisType>& order);
 
 public:
 
-    c_RotationMatrix(amrex::Vector<T> angles, 
-                   AngleType type = AngleType::Degrees) :
-    _angles(type == AngleType::Degrees ? angles * M_PI / 180 : angles),
-    _rotX(3, amrex::Vector<T>(3)),
-    _rotY(3, amrex::Vector<T>(3)),
-    _rotZ(3, amrex::Vector<T>(3))
-    {
-        static_assert(angles.size() == 3, "Angles vector must be of size 3");
-        Define_RotationMatrices();
+    c_RotationMatrix(amrex::Vector<T> angles,
+                     AngleType type); //default type AngleType::Degrees
+
+    amrex::Vector<T> Get_RotationAngles() const { 
+        return _angles;
     }
 
-    void Update_RotationAngles(amrex::Vector<T> new_angles) 
-    {
-        static_assert(new_angles.size() == 3, "Angles vector must be of size 3");
-        _angles = new_angles;
+    void Update_RotationMatrices(amrex::Vector<T>& new_angles);
 
-        Define_RotationMatrices();
-    }
-
-    amrex::Vector<T> RotateVector(const amrex::Vector<T>& v1, 
-                                  const std::vector<AxisType>& order = {})
-    {
-        assert(v1.size() == 3);
-
-        if (order.empty()) return; 
-
-        ApplyRotationOrder(order, v1);
-    }
+    template<typename ContainerType>
+    void RotateContainer(const ContainerType& v1,
+                         const std::vector<AxisType>& order); 
+                      //default order = {} 
 
 };
+#endif
